+++ resources/assets/js/components/projects/single.vue
@@ -12,6 +12,9 @@
               Show All Members
             </a>
           </li>
+          <li @click="showGithubRepoModal" class="px-4 py-2 hover:bg-grey-light cursor-pointer">
+            Connect Github Repository
+          </li>
           <li class="px-4 py-2 hover:bg-grey-light cursor-pointer">
             Delete
           </li>
@@ -19,8 +22,9 @@
       </div>
     </div>
 
-    <!-- Add Member Form -->
-    <addMemberForm v-if="addMemberFormShown" @close="closeAddMemberForm" resourceType="project" :resource="project" @addMember="addMember"></addMemberForm>
+    <add-member-form v-if="addMemberFormShown" @close="closeAddMemberForm" resourceType="project" :resource="project" @addMember="addMember"></add-member-form>
+
+    <show-github-repo entityType="project" :entityId="project.id" v-if="githubRepoModalShown" @close-github-repo-modal="closeGithubRepoModal"></show-github-repo>
 
     <div class="h-16 flex flex-row justify-center items-center px-2">
       <span @click="showAddMemberForm" class="bg-white shadow w-8 h-8 rounded-full text-teal hover:cursor-pointer text-center p-2">
@@ -55,6 +59,7 @@ import eventBoard from './../partials/eventBoard.vue'
 import fileBoard from './../partials/fileBoard.vue'
 import activity from './../partials/activity.vue'
 import addMemberForm from './../partials/addMemberForm.vue'
+import showGithubRepo from './../partials/showGithubRepo.vue'
 import membersListModal from './../partials/membersListModal.vue'
 import tabMenu from './../partials/tabMenu.vue'
 import { faPlus } from '@fortawesome/free-solid-svg-icons/faPlus'
@@ -71,12 +76,14 @@ export default {
     addMemberForm,
     membersListModal,
     tabMenu,
+    showGithubRepo
   },
   props: ['project'],
   data: () => ({
     addMemberFormShown: false,
     active: 'tasks',
     dropdownMenuShown: false,
+    githubRepoModalShown: false,
     membersListModalShown: false,
     faPlus,
     faCog
@@ -122,6 +129,12 @@ export default {
     },
     closeDropdownMenu () {
       this.dropdownMenuShown = false
+    },
+    showGithubRepoModal () {
+      this.githubRepoModalShown = true
+    },
+    closeGithubRepoModal () {
+      this.githubRepoModalShown = false
     }
   }
 }