+++ resources/assets/js/components/teams/single.vue
@@ -22,6 +22,8 @@
     <!-- Add Member Form -->
     <addMemberForm v-if="addMemberFormShown" @close="closeAddMemberForm" resourceType="team" :resource="team" @addMember="addMember"></addMemberForm>
 
+    <show-github-repo entityType="team" :entityId="team.id" v-if="githubRepoModalShown" @close-github-repo-modal="closeGithubRepoModal"></show-github-repo>
+
     <div class="h-16 flex flex-row justify-center items-center px-2">
       <span @click="showAddMemberForm" class="bg-white shadow w-8 h-8 rounded-full text-teal hover:cursor-pointer text-center p-2">
         <font-awesome-icon :icon="faPlus"></font-awesome-icon>
@@ -56,6 +58,7 @@ import eventBoard from './../partials/eventBoard.vue'
 import fileBoard from './../partials/fileBoard.vue'
 import activity from './../partials/activity.vue'
 import addMemberForm from './../partials/addMemberForm.vue'
+import showGithubRepo from './../partials/showGithubRepo.vue'
 import membersListModal from './../partials/membersListModal.vue'
 import tabMenu from './../partials/tabMenu.vue'
 import { faPlus } from '@fortawesome/free-solid-svg-icons/faPlus'
@@ -72,12 +75,14 @@ export default {
     addMemberForm,
     membersListModal,
     tabMenu,
+    showGithubRepo
   },
   props: ['team'],
   data: () => ({
     addMemberFormShown: false,
     active: 'tasks',
     dropdownMenuShown: false,
+    githubRepoModalShown: false,
     membersListModalShown: false,
     faPlus,
     faCog
@@ -123,6 +128,12 @@ export default {
     },
     closeDropdownMenu () {
       this.dropdownMenuShown = false
+    },
+    showGithubRepoModal () {
+      this.githubRepoModalShown = true
+    },
+    closeGithubRepoModal () {
+      this.githubRepoModalShown = false
     }
   }
 }