+++ CHANGELOG.md
@@ -3,13 +3,13 @@
 ## v0.9 (Yet to release)
 
 - Added direct message to a user functionality
+- Added mentioning user functionality in comment & message
 - Added endpoint to list all tags
 - Added endpoint to create categories
 - Added file upload option in file section
 - Added endpoint to check if username exists
 - Added finnish translation
 - Added simplified Chinese translation
-- Add telescope installation to wiki
 
 ## v0.8 (2018-10-18)